examples/rag-langchain-python/requirements.txt
blindfold-sdk>=1.3.0
langchain-blindfold>=0.1.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
faiss-cpu>=1.9.0
python-dotenv>=1.0.0
